requirements.txt
google-genai>=1.0.0
openai>=1.0.0
yfinance>=0.2.40
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
mplfinance>=0.12.10b0
